+++ src/pages/NftItemPage.vue
@@ -1,10 +1,13 @@
 <script lang="ts" setup>
-import { Loader, ErrorMessage } from '@/common'
+import { Loader, ErrorMessage, AppButton } from '@/common'
 import { NftDetails, SaleHistory } from '@/pages/nft-item-page'
 
 import { ErrorHandler, getBookById } from '@/helpers'
-import { ref } from 'vue'
+import { ref, computed } from 'vue'
 import { BookRecord } from '@/records'
+import { WINDOW_BREAKPOINTS } from '@/enums'
+import { useWindowSize } from '@vueuse/core'
+import { useI18n } from 'vue-i18n'
 
 const props = defineProps<{
   id: string
@@ -14,6 +17,9 @@ const isLoaded = ref(false)
 const isLoadFailed = ref(false)
 
 const book = ref<BookRecord | undefined>()
+
+const { width } = useWindowSize()
+const { t } = useI18n()
 
 const init = async () => {
   try {
@@ -25,6 +31,10 @@ const init = async () => {
   }
   isLoaded.value = true
 }
+
+const buttonLinkText = computed(() =>
+  width.value >= WINDOW_BREAKPOINTS.small ? t('nft-item-page.edit-button') : '',
+)
 
 init()
 </script>
@@ -57,6 +67,19 @@ init()
     <template v-else>
       <loader />
     </template>
+
+    <mounted-teleport to="#app-navbar__right-buttons">
+      <app-button
+        class="nft-item-page__link-button"
+        size="small"
+        :icon-left="$icons.edit"
+        :text="buttonLinkText"
+        :route="{
+          name: $routes.nftItemEdit,
+          params: { id: props.id },
+        }"
+      />
+    </mounted-teleport>
   </div>
 </template>
 
@@ -116,5 +139,16 @@ init()
 
 .nft-item-page__tabs {
   margin-bottom: toRem(40);
+}
+
+.nft-item-page__link-button {
+  width: toRem(180);
+  order: -1;
+
+  @include respond-to(small) {
+    width: toRem(54);
+    height: toRem(54);
+    order: 1;
+  }
 }
 </style>